Use babel instead of locale
Before, the locale was change globally with locale.setlocale in order to
format numbers and dates, but it also changed the locale for the python
program importing tkfilebrowser, which can have undesired effects such as
writing numbers with locale format in a file, which makes it unusable with
other softwares. With babel, the numbers and dates can be formatted with
the default locale without changing python locale.